split report formatters by output mode
